requirements.txt
pymongo
python-dotenv
paho-mqtt